updates erc20 example
